[0.1.0B] - Debug font
 * Wrote a python script that converts black and white images into chars
 * Loaded debug font texture
 * A few other changes

+++ Kondion_N/src/KondionMath.cpp
@@ -119,8 +119,8 @@ void printWorld() {
       running = false;
     }
   }
-  
-  printf("E: %u %u %u\n", kondismol_data[0], kondismol_data[1], kondismol_data[2]);
+
+  printf("E: %u %u %u\n", uint8_t(kondismol_data[0]), uint8_t(kondismol_data[1]), uint8_t(kondismol_data[2]));
 
   //delete[] stack
 }
